refactor utils.py from being the kitchen sink
* more logical separation of network/file io and exceptions
* add a handful of TODOs for minor things
